ver1/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
pillow>=8.3.1
tqdm>=4.62.2
matplotlib>=3.4.3
pandas>=1.3.3
scikit-learn>=0.24.2 